+++ src/metworkpy/network/density.py
@@ -2,17 +2,25 @@
 
 # Standard Library Imports
 from __future__ import annotations
-from typing import Callable, Hashable, Union, Literal, Tuple, Optional
+
+from typing import (
+    Callable,
+    Hashable,
+    Iterable,
+    Literal,
+    Optional,
+    Tuple,
+    Union,
+)
 from warnings import warn
 
 # External Imports
-import cobra  # type:ignore     # Cobra doesn't have py.typed marker
-from joblib import Parallel, delayed
+import cobra
 import networkx as nx
 import numpy as np
 import pandas as pd
+from joblib import Parallel, delayed
 from scipy import stats
-
 
 # Local Imports
 from metworkpy.network.neighborhoods import (
@@ -21,13 +29,13 @@ from metworkpy.network.neighborhoods import (
 )
 from metworkpy.utils.translate import get_reaction_to_gene_translation_dict
 
-
 # region Main Functions
 
 
 def node_target_density(
     network: nx.Graph | nx.DiGraph,
     targets: list[Hashable] | dict[Hashable, float | int] | pd.Series,
+    nodes: Optional[Iterable[Hashable]] = None,
     radius: int = 3,
     node_filter: Optional[
         Union[Callable[[Hashable], bool], set[Hashable]]
@@ -50,12 +58,15 @@ def node_target_density(
         Series keyed by nodes in the network which can specify a target
         weight (such as multiple targets for a single node). If a dict or
         Series, values should be ints or floats.
+    nodes : iterable of hashable, optional
+        Subset of nodes to find the density for, if not provided defaults
+        to all of the nodes in the network
     radius : int
         Radius to use for finding density. Specifies how far out from a
         given node targets are counted towards density. A radius of 0
         only counts the single node, and so will just return the
         `targets` values back unchanged. Default value of 3.
-    node_filter : Callable of node id to bool or set of node id, optional
+    node_filter : Callable of node id to bool, or set of node id, optional
         Filter nodes in the network to consider when calculating density.
         If a Callable, should take node ids as the only argument and return
         a bool, if True the node will be considered in the density,
@@ -101,7 +112,10 @@ def node_target_density(
         def filter_fn(_: Hashable) -> bool:
             return True
 
-    results_series = pd.Series(np.nan, index=pd.Index(network.nodes))
+    if nodes is None:
+        nodes = network.nodes
+
+    results_series = pd.Series(np.nan, index=pd.Index(nodes))
     for node, density in Parallel(n_jobs=processes, return_as="generator")(
         delayed(_node_density_worker)(
             node,
@@ -110,7 +124,7 @@ def node_target_density(
             node_filter=filter_fn,
             radius=radius,
         )
-        for node in network.nodes
+        for node in nodes
     ):
         results_series[node] = density
     return results_series
@@ -120,6 +134,7 @@ def gene_target_density(
     metabolic_network: Union[nx.Graph, nx.DiGraph],
     metabolic_model: cobra.Model,
     gene_targets: Union[pd.Series, list, dict],
+    nodes: Optional[Iterable[Hashable]] = None,
     radius: int = 3,
     essential: bool = False,
     processes: Optional[int] = None,
@@ -142,6 +157,9 @@ def gene_target_density(
         have equal weight. If a dict, should be keyed by gene id, with values
         corresponding to weight. If a pd.Series, should be indexed by gene id,
         with values corresponding to weight.
+    nodes : iterable of hashable, optional
+        Subset of nodes to find the density for, if not provided defaults
+        to all of the nodes in the network
     radius : int, default=3
         The radius to use for finding density, specifies how far out from
         a given node targets are counted towards density. A radius of 0 only
@@ -170,11 +188,13 @@ def gene_target_density(
             f"Metabolic network must be a networkx Graph but received a "
             f"{type(metabolic_network)}"
         )
+    if nodes is None:
+        nodes = metabolic_network.nodes
     if isinstance(gene_targets, list):
         gene_targets = pd.Series(1, index=pd.Index(gene_targets))
     elif isinstance(gene_targets, dict):
         gene_targets = pd.Series(gene_targets)
-    density_series = pd.Series(np.nan, index=pd.Index(metabolic_network.nodes))
+    density_series = pd.Series(np.nan, index=pd.Index(nodes))
     rxn_to_gene_set_dict = get_reaction_to_gene_translation_dict(
         model=metabolic_model, essential=essential
     )
@@ -186,7 +206,7 @@ def gene_target_density(
             radius=radius,
             rxn_to_gene_set_dict=rxn_to_gene_set_dict,
         )
-        for node in metabolic_network.nodes
+        for node in nodes
     ):
         density_series[node] = density
     return density_series
@@ -196,6 +216,7 @@ def gene_target_enrichment(
     metabolic_network: Union[nx.Graph, nx.DiGraph],
     metabolic_model: cobra.Model,
     gene_targets: Union[set[str], list[str]],
+    nodes: Optional[Iterable[Hashable]] = None,
     metric: Literal["odds-ratio", "p-value"] = "p-value",
     alternative: Literal["two-sided", "less", "greater"] = "greater",
     radius: int = 3,
@@ -218,6 +239,9 @@ def gene_target_enrichment(
         Targeted genes associated with reactions in the
         metabolic network. Result will be the enrichment in these targeted
         genes in a neighborhood of each reaction in the network
+    nodes : iterable of hashable, optional
+        Subset of nodes to find the enrichment for, if not provided defaults
+        to all of the nodes in the network
     metric : "odds-ratio" or "p-value", default="p-value"
         The enrichment metric to return in the Series, either the odds-ratio
         or the p-value (default) of the Fisher's exact test used to
@@ -261,18 +285,18 @@ def gene_target_enrichment(
             f"Gene targets must be a list or a set but received a "
             f"{type(gene_targets)}"
         )
+    if nodes is None:
+        nodes = metabolic_network.nodes
     # Filter the gene targets for only those in the model
     gene_targets &= set(metabolic_model.genes.list_attr("id"))
     if len(gene_targets) < 1:
         warn("No targeted genes, p-values all 1.0, odds-ratio all 0.0")
         if metric == "p-value":
-            return pd.Series(1.0, index=pd.Index(metabolic_network.nodes))
+            return pd.Series(1.0, index=pd.Index(nodes))
         elif metric == "odds-ratio":
-            return pd.Series(0.0, index=pd.Index(metabolic_network.nodes))
+            return pd.Series(0.0, index=pd.Index(nodes))
     total_genes = len(metabolic_model.genes.list_attr("id"))
-    enrichment_series = pd.Series(
-        np.nan, index=pd.Index(metabolic_network.nodes)
-    )
+    enrichment_series = pd.Series(np.nan, index=pd.Index(nodes))
     rxn_to_gene_set_dict = get_reaction_to_gene_translation_dict(
         model=metabolic_model, essential=essential
     )
@@ -287,7 +311,7 @@ def gene_target_enrichment(
             rxn_to_gene_set_dict=rxn_to_gene_set_dict,
             total_genes=total_genes,
         )
-        for node in metabolic_network.nodes
+        for node in nodes
     ):
         enrichment_series[node] = pval if metric == "p-value" else odds
     return enrichment_series